remove Button component element.js
