--- src/components/Button/elements.js
@@ -1,31 +1,0 @@
-import styled, { css } from 'styled-components'
-import { Button as RebassButton } from 'rebass'
-
-const baseStyles = css`
-  text-transform: uppercase;
-  text-decoration: none;
-  min-height: 50px;
-  border-radius: 5px;
-  font-size: 0.8em;
-  display: flex;
-  flex: none;
-  align-self: center;
-  align-items: center;
-  justify-content: center;
-  font-weight: 600;
-  white-space: nowrap;
-  word-break: keep-all;
-  cursor: ${props => (props.disabled ? 'not-allowed' : 'pointer')};
-`
-
-export const BaseButton = styled(RebassButton)`
-  ${baseStyles}
-`
-
-export const Label = styled.span`
-  display: block;
-  flex: 0 0 auto;
-  line-height: inherit;
-  color: inherit;
-  align-self: center;
-`
