+++ src/components/Button/index.tsx
@@ -1,13 +1,37 @@
 import React from 'react'
-
 import Icon from 'src/components/Icons'
+import styled, { css } from 'styled-components'
 import {
-  ColorProps,
-  SpaceProps,
-  WidthProps,
-  ButtonStyleProps,
-} from 'styled-system'
-import { BaseButton, Label } from './elements'
+  Button as RebassButton,
+  ButtonProps as RebassButtonProps,
+} from 'rebass'
+
+const baseStyles = css`
+  min-height: 50px;
+  border-radius: 5px;
+  display: flex;
+  flex: none;
+  align-self: center;
+  align-items: center;
+  justify-content: center;
+  white-space: nowrap;
+  word-break: keep-all;
+  /* cursor: ${props => (props.disabled ? 'not-allowed' : 'pointer')}; */
+`
+
+export const BaseButton = styled(RebassButton)`
+  ${baseStyles}
+`
+
+export const Label = styled.span`
+  text-transform: uppercase;
+  text-decoration: none;
+  display: block;
+  flex: 0 0 auto;
+  line-height: inherit;
+  color: inherit;
+  align-self: center;
+`
 
 // extend to allow any default button props (e.g. onClick) to also be passed
 export interface IBtnProps extends React.ButtonHTMLAttributes<HTMLElement> {
@@ -15,27 +39,17 @@ export interface IBtnProps extends React.ButtonHTMLAttributes<HTMLElement> {
 }
 
 interface IDisabledProp {
-  boolean?: false
+  disabled?: boolean
 }
+
+type BtnProps = IBtnProps & RebassButtonProps & IDisabledProp
 
 const StyledButton = ({ children, ...props }) => (
   <BaseButton {...props}>{children}</BaseButton>
 )
 
-StyledButton.defaultProps = {
-  className: 'button',
-  variant: 'primary',
-}
-
-type BtnProps = IBtnProps &
-  SpaceProps &
-  WidthProps &
-  ButtonStyleProps &
-  ColorProps &
-  IDisabledProp
-
 export const Button = (props: BtnProps) => (
-  <StyledButton px={3} {...props}>
+  <StyledButton className="button" variant="primary" px={3} {...props}>
     {props.icon && <Icon glyph={props.icon} />}
     <Label>{props.children}</Label>
   </StyledButton>
